% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal.R
\name{Correlation}
\alias{Correlation}
\title{Correlation}
\usage{
Correlation(vectorx, vectory)
}
\arguments{
\item{vectorx}{1st vector parameter}

\item{vectory}{2nd vector parameter}
}
\value{
returns the correlation of the two vectors
}
\description{
Correlation
}
