% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_matrix.R
\name{correlation_matrix}
\alias{correlation_matrix}
\title{Correlation matrix function}
\usage{
correlation_matrix(X)
}
\arguments{
\item{X}{}
}
\value{
a matrix (the sample correlation matrix of X)
}
\description{
Takes a data array and returns the sample correlation matrix
}
\examples{
X = [1 2 3
     4 5 6]; correlation_matrix(X)

}
