% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Turn in Functions.R
\name{CorrelationMatrix}
\alias{CorrelationMatrix}
\title{Correlation Matrix Function}
\usage{
CorrelationMatrix(matrix)
}
\arguments{
\item{matrix}{matrix of values used to calculate correlation, where columns represent datasets and rows represent n datapoints for each dataset.}
}
\value{
a matrix of correlation values for the data
}
\description{
Correlation Matrix Function
}
