% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrmatrix.R
\name{corrmat}
\alias{corrmat}
\title{Sample correlation function}
\usage{
corrmat(X)
}
\arguments{
\item{X}{}
}
\value{
A matrix of sample correlations
}
\description{
This function returns a matrix of sample correlations by using the input.
}
\examples{
X = matrix(1:9, 3, 3), corrmat(X)
}
