% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal.R
\name{Covariance}
\alias{Covariance}
\title{Covariance}
\usage{
Covariance(vectorx, vectory)
}
\arguments{
\item{vectorx}{1st vector parameter}

\item{vectory}{2nd vector parameter}
}
\value{
returns the covariance of the two vectors
}
\description{
Covariance
}
