% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Turn in Functions.R
\name{CovarianceMatrix}
\alias{CovarianceMatrix}
\title{Covariance Matrix Function}
\usage{
CovarianceMatrix(matrix)
}
\arguments{
\item{matrix}{matrix of vectors, with the number of vectors represented by columns, and the amount of data represented as rows}
}
\value{
a matrix of the covariances of all the vectors
}
\description{
Covariance Matrix Function
}
\examples{
a matrix 4x3 matrix will return a 4x4 zero matrix
}
