% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covmatrix.R
\name{covmat}
\alias{covmat}
\title{Sample variances and covariances function}
\usage{
covmat(X)
}
\arguments{
\item{X}{}
}
\value{
A matrix of variances and covariances
}
\description{
The function constructs a matrix of variances and covariances by using the input.
}
\examples{
X = matrix(1:9, 3, 3), covmat(X)
}
