% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_vector.R
\name{mean_vector}
\alias{mean_vector}
\title{Mean vector function}
\usage{
mean_vector(X)
}
\arguments{
\item{X}{}
}
\value{
a vector (the mean vector of X)
}
\description{
Takes a data array and returns the sample mean vector
}
\examples{
X = [1 2 3
     4 5 6]; mean_vector(X)

}
