% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanvector.R
\name{meanvector}
\alias{meanvector}
\title{Mean vector function}
\usage{
meanvector(X)
}
\arguments{
\item{X}{}
}
\value{
A vector of sample means
}
\description{
The function takes the input and constructs a vector composed of each sample mean
}
\examples{
X = 1:30, meanvector(X)
}
