% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycenter.R
\name{mycenter}
\alias{mycenter}
\title{Subtract column means from columns of a matrix}
\usage{
mycenter(X)
}
\arguments{
\item{X}{a data matrix}
}
\value{
a centered matrix.
}
\description{
Subtract the mean from the columns.
}
\details{
This utility function subtracts the mean from each of the columns.
}
