% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycorrelation.R
\name{mycorrelation}
\alias{mycorrelation}
\title{Compute correlation matrix}
\usage{
mycorrelation(X)
}
\arguments{
\item{X}{data matrix}
}
\value{
a correlation matrix
}
\description{
Creates a correlation matrix.
}
\details{
This function creats a correlation matrix from a data matrix.
}
\examples{
X = matrix(runif(4*4), 4, 4); mycorrelation(X)

}
