% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycovariance.R
\name{mycovariance}
\alias{mycovariance}
\title{Compute covariance matrix}
\usage{
mycovariance(X)
}
\arguments{
\item{X}{data matrix}
}
\value{
a covariance matrix
}
\description{
Creates a covariance matrix
}
\details{
This function creates a covariance matrix from a data matrix.
}
\examples{
X = matrix(runif(4*4), 4, 4); mycovariance(X)
}
