% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymean.R
\name{mymean}
\alias{mymean}
\title{Calculate vector mean}
\usage{
mymean(x)
}
\arguments{
\item{x}{a vector}
}
\value{
mean of the vector
}
\description{
Calculate the mean of a vector.
}
\details{
This utility function calculates the mean of a vector.
}
\examples{

x=rnorm(10); mymean(x)

}
