% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymeanvector.R
\name{mymeanvector}
\alias{mymeanvector}
\title{Calculate column means of a matrix}
\usage{
mymeanvector(X)
}
\arguments{
\item{X}{a data matrix}
}
\value{
a mean vector
}
\description{
Calculate a mean vector.
}
\details{
This function calculates a mean vector.
}
\examples{
X = matrix(runif(4*4), 4, 4); mymeanvector(X)
}
