% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mySampleCorrelationMatrix.R
\name{mySampleCorrelationMatrix}
\alias{mySampleCorrelationMatrix}
\title{mySampleCorrelationMatrix}
\usage{
mySampleCorrelationMatrix(m)
}
\arguments{
\item{m}{A matrix}
}
\value{
A matrix
}
\description{
Calculates the Sample Correlation Matrix
}
