% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mySampleCovarienceMatrix.R
\name{mySampleCovarienceMatrix}
\alias{mySampleCovarienceMatrix}
\title{mySampleCovariepceMatrix}
\usage{
mySampleCovarienceMatrix(m)
}
\arguments{
\item{m}{A matrix}
}
\description{
Calculates a biased sample covariepce matrix
}
