% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mySampleMeanVector.R
\name{mySampleMeanVector}
\alias{mySampleMeanVector}
\title{mySampleMeanVector}
\usage{
mySampleMeanVector(x)
}
\arguments{
\item{x}{A matrix}
}
\value{
A vector
}
\description{
Calculate the sample mean vector from a matrix
}
