% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samcor.R
\name{samcor}
\alias{samcor}
\title{Sample Correlation of Vectors}
\usage{
samcor(x)
}
\arguments{
\item{x}{a matrix}
}
\value{
a correlation matrix
}
\description{
Takes a matrix and outputs the sample correlation matrix of the variables represented
}
\examples{
samcor(paper.df)
}
