% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samcov.R
\name{samcov}
\alias{samcov}
\title{Sample Covariance of Vectors}
\usage{
samcov(x)
}
\arguments{
\item{x}{a matrix}
}
\value{
a covariance matrix
}
\description{
Takes a matrix and outputs the sample covariance matrix of the variables represented.
}
\examples{
samcov(paper.df)
}
