% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sammean.R
\name{sammean}
\alias{sammean}
\title{Sample Mean Vector}
\usage{
sammean(x)
}
\arguments{
\item{x}{a matrix}
}
\value{
A vector of sample means of all variables
}
\description{
Takes a matrix and outputs sample means for all columns in vector form
}
\examples{
sammean(x)
}
