% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampcor.R
\name{sampcor}
\alias{sampcor}
\title{Sample correlation matrix}
\usage{
sampcor(x, varnames = NULL)
}
\arguments{
\item{x}{data set}

\item{varnames}{list of names}
}
\value{

}
\description{
Returns a correlation matrix for a data sample
}
\details{
Accepts a data set and creates a correlation matrix using the sampcov function and the definition of the
correlation matrix
}
\examples{
sampcor(data, c("WEIGHT", "LENGTH", "DENSITY"))
}
