% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampcov.R
\name{sampcov}
\alias{sampcov}
\title{Sample Covariance (biased)}
\usage{
sampcov(x, varnames = NULL)
}
\arguments{
\item{x}{a data set}

\item{varnames}{a list of variable names}
}
\value{

}
\description{
Returns a (biased) covariance matrix for a data matrix
}
\details{
Accepts a data set and creates a $p\times p$ covariance matrix, where p is the number of variables.
The function also accepts variable names that will be used to lable the columns of the matrix.
These results are biased, as they use $\frac{1}{n}$ in the definition of the variance, rather than $\frac{1}{n-1}$.
}
\examples{
sampcov(data, c("WEIGHT", "LENGTH", "DENSITY"))
}
