% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleBiasedCovarianceMatrix.R
\name{SampleBiasedCovarianceVariance}
\alias{SampleBiasedCovarianceVariance}
\title{Sample (Biased) Covariance with Variance}
\usage{
SampleBiasedCovarianceVariance(x)
}
\arguments{
\item{x}{A Matrix, where each variable represents a column}
}
\value{
A vector of the sample means
}
\description{
Returns a matrix of the covariance withvariance (baised) of the given matrix.
}
\details{
Given a matrix of data, an algorithm is employed to calculate the
sample covariance and variance (biased) matrix. The variance values
are located on the diagonal of the matrix, x11,x22...xpp; while the
covariance values are located in the other positions.
Note: Uses SampleMeanVector function.
}
\examples{
myMat = c(3,4,2,6,8,2,5, 5,5.5,4,7,10,5,7.5)
dim(myMat) = c(7,2) # Creates a 2x7 matrix
mySampleCovarianceVarianceVariable = SampleBiasedCovarianceVariance(myMat)
}
