% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleCor.R
\name{SampleCor}
\alias{SampleCor}
\title{Sample Correlation Function}
\usage{
SampleCor(x)
}
\arguments{
\item{x}{a matrix}
}
\value{
a correlation matrix
}
\description{
Finds (biased) sample correlation
}
\details{
This function helps us obtain the correlation matrix
}
