% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleCov.R
\name{SampleCov}
\alias{SampleCov}
\title{Sample Covariance Function}
\usage{
SampleCov(x)
}
\arguments{
\item{x}{a matrix}
}
\value{
a covariance matrix
}
\description{
Finds biased sample covariance
}
\details{
This function helps us obtain the covariance matrix
}
