% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleMeans.R
\name{SampleMeans}
\alias{SampleMeans}
\title{Sample Means Function}
\usage{
SampleMeans(x)
}
\arguments{
\item{x}{a matrix}
}
\value{
a vector of sample means
}
\description{
Finds sample mean for a vector of data
}
\details{
This function helps us obtain the sample mean for each column of data in matrix
}
