% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleMeanVector.R
\name{SampleMeanVector}
\alias{SampleMeanVector}
\title{Sample Mean Vector}
\usage{
SampleMeanVector(x)
}
\arguments{
\item{x}{A Matrix, where each variable represents a column}
}
\value{
A vector of the sample means
}
\description{
Returns the mean vector of the given matrix.
}
\details{
Given a matrix of data, an algorithm is employed to calculate the
sample mean vector.
}
\examples{
myMat = c(3,4,2,6,8,2,5, 5,5.5,4,7,10,5,7.5)
dim(myMat) = c(7,2) # Creates a 2x7 matrix
mySampleMeanVariable = SampleMeanVector(myMat)
}
