% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampmean.R
\name{sampmean}
\alias{sampmean}
\title{Sample mean}
\usage{
sampmean(x)
}
\arguments{
\item{x}{a data matrix}
}
\value{

}
\description{
Returns a vector of the sample mean
}
\details{
Accepts a data set and takes the mean of each variable and then outputs a
vector containing each of those means
}
\examples{
sampmean(data)
}
