% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Turn in Functions.R
\name{VectorMean}
\alias{VectorMean}
\title{Vector Mean}
\usage{
VectorMean(x)
}
\arguments{
\item{x}{a vector object}
}
\value{
The meean of the vector object passed as a parameter
}
\description{
Vector Mean
}
\examples{
The mean of a sample vector [1,1,1,1] = 1. The mean of a vector [1,2,3,4,5] = 3


}
